requirements.txt
antlr4-python3-runtime==4.9.3
colorama==0.4.5
hydra-core==1.2.0
omegaconf==2.2.2
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
tqdm==4.64.0
